create UnpublishedWorkspaceSpec
fix

remove unpublish workspace

lower Xmx
