+++ automation/build.sbt
@@ -63,7 +63,7 @@ Global / concurrentRestrictions := Seq(Tags.limit(Tags.ForkedTestGroup, 5))
 /**
   * Forked JVM options
   */
-Test / javaOptions ++= Seq("-Xmx4G")
+Test / javaOptions ++= Seq("-Xmx2G")
 
 /**
  * copy system properties to forked JVM
